remove map names file
